Script for job monitoring and handling in Totango
This script is a part of our efforts to enhance data management and monitoring in Totango. It is designed to fetch the details of all data jobs, whether they are running on SFTP or non-SFTP, that are currently operating in Totango. The aim of the script is to facilitate the creation of a day-to-day load monitoring dashboard, enabling us to take quick remedial action in case any job fails.